# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  utils.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:4000
info:
  contact:
    email: davidhero125@gmail.com
    name: API Support
  description: Invoice management and payment platform API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Stellance API
  version: "1.0"
paths:
...
  /wallet:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create a Stellar wallet
      tags:
      - wallet
  /wallet/{id}:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get wallet by ID
      tags:
      - wallet
  /wallet/{id}/export:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Export wallet keys
      tags:
      - wallet
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"